Change assert about breakpoint location
In the context of a breakpoint placed in a header, gdb might change its
line number since it is not reachable from any of the loaded object
files. Later, when a new object is loaded that can reach it, gdb will
move the breakpoint. modifiedFlag will be set in this case.