Remove asserts and modified flag to HandleNewBreakpoint
The flag was used only to perform these asserts and they fail in some
cases. This commit is an improvement in general, but there might be
another bug relating to breakpoints. The sign might need "unplacing"
if the breakpoint was modified to a new location. This potential issue
is present in previous commits and I haven't observed it in practice.